feat(Books w/ Clean-Code): Mock server-call to make test pass
