+++ src/bookCleanCode/hooks/useBooks.ts
@@ -1,12 +1,12 @@
 import { useState } from 'react';
-import { useBooksAdapterFactory } from '../adapter/useBooksAdapter';
+import { booksAdapterFactory } from '../adapter/booksAdapter';
 import type { BookServerModel } from '../interfaces';
 import { createBooks } from '../services/BooksService';
 
 export const useBooks = () => {
 	const [books, setBooks] = useState<BookServerModel[]>([]);
 	const [error, setError] = useState<Error>();
-	const createBookAdapter = useBooksAdapterFactory();
+	const createBookAdapter = booksAdapterFactory();
 	const loadAllBooksFromServer = createBookAdapter();
 
 	const loadAllBooks = async () => {
